Move cublas helper code from test into cuda_support lib
